Change variable type from int to float
Changed variable type from int to float for better precision.

+++ exercicio3.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() { 
-    int número;
+    float número;
 
     printf ("Digite um número: ");
     scanf ("%i", &número);
